Cpu+Debugger: moved the debugger line out of `psx-core`
Now, the emulator will provide a programmable interface to control the
debugger and have internal access to the CPU component that are useful
to have to build a debugger.

Now, the UI parts of the debugger are moved outside into
`src/debugger.rs` which communicates with the CPU debugger and requests
changes from it. Breakpoints are triggered and returned as part of the
new `CpuState`, which tells why the CPU stopped now, `Normal` means it
it only stopped after the certain amount of clocks finished.
