+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 default = ["debugger"]
-debugger = ["psx-core/debugger"]
+debugger = ["psx-core/debugger", "dep:rustyline"]
 
 [dependencies]
 psx-core = { path = "./psx-core", version = "*" }
@@ -19,6 +19,8 @@ winit = "0.28"
 vulkano-win = "0.33"
 ringbuf = { version = "0.3.1", default-features = false, features = ["alloc"] }
 cpal = { version = "0.15", default-features = false }
+
+rustyline = { version = "11.0.0", default-features = false, optional = true }
 
 [workspace]
 members = [
